requirements.txt
Flask==3.1.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-WTF==1.2.2
Flask-Compress==1.15
Flask-Limiter==3.7.0
Flask-Mail==0.10.0
SQLAlchemy==2.0.46
psycopg2-binary==2.9.9
redis==5.0.1
python-dotenv==1.2.1
python-dateutil==2.9.0.post0
Werkzeug==3.1.3
WTForms==3.2.2
email-validator==2.2.0
requests==2.32.5
Pillow==10.4.0
openpyxl==3.1.5
gunicorn==23.0.0
